% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\name{db.connection}
\alias{db.connection}
\title{Get the DBIConnection connection object by config name.}
\usage{
db.connection(con.name)
}
\arguments{
\item{con.name}{connection names are either defined in database.yml or registered
at runtime using register.connections(). If connection configuration exists,
but the DB connection has not yet been established,
calling this function will also try to connect to the database.}
}
\description{
Get the DBIConnection connection object by config name.
}
\seealso{
list.connections, register.connections
}
