% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\name{db.disconnect}
\alias{db.disconnect}
\title{Disconnect database connections.}
\usage{
db.disconnect(con.name = NA)
}
\arguments{
\item{con.name}{specify connection name as defined (in e.g. database.yml).
If it is not NA, then a connection is closed by the given name. Otherwise
all open database connections are closed.}
}
\description{
Disconnect database connections.
}
\seealso{
list.connections
}
