% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\name{db}
\alias{db}
\title{Read from a Database (currently only PostgreSQL) connection.}
\usage{
db(con.names, sql, params = c(), cores = 4)
}
\arguments{
\item{con.names}{a vector of connection names as defined in database.yml (or
using custom connection definitions). If more than one connection names is
given then the same query is performed on all connections in parallel. This
is particularly useful for analytical queries on sharded setup. For example:

  shards <- paste('shard', 1:16, sep='')
  db(shards, 'select count(*) from events'))

will run in parallel on all 16 shards.

TODO: exemplify the params binding feature}
}
\description{
Read from a Database (currently only PostgreSQL) connection.
}
