% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\docType{data}
\name{.RportRuntimeEnv}
\alias{.RportRuntimeEnv}
\title{The structure of the store variable living in .RportRuntimeEnv is like this:}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.RportRuntimeEnv
}
\description{
rport.store <- list(
  db.connections::shard1="PgConnection",
  db.connections::shard2="PgConnection",
  ...
  db.config=list(shard1=list(dbname=shard1), shard2=list(dbname=shard2)),
  db.driver="DbDriver",
)
}
\keyword{datasets}
