% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\name{list.connections}
\alias{list.connections}
\title{List open DB connections.}
\usage{
list.connections()
}
\description{
This only lists all connections that have been open. No new connections will be opened by this call.
}
