% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_copy.R
\name{pg.copy}
\alias{pg.copy}
\title{A lightweight wrapper around postgresqlCopyInDataframe that allows for
custom field lists and temporary tables. It only allows numeric, character,
integer, date/time and bool types. It's more useful than dbWriteTable because
it allows SQL-copying into tables that have fields other than the data.frame.
Also using pg.copy you can SQL-copy in transactions with temp tables.}
\usage{
pg.copy(con, tbl.name, DT)
}
\arguments{
\item{con}{PostgreSQLConnection}

\item{tbl.name}{character}

\item{tbl.name}{PostgreSQLConnection}
}
\description{
A lightweight wrapper around postgresqlCopyInDataframe that allows for
custom field lists and temporary tables. It only allows numeric, character,
integer, date/time and bool types. It's more useful than dbWriteTable because
it allows SQL-copying into tables that have fields other than the data.frame.
Also using pg.copy you can SQL-copy in transactions with temp tables.
}
