\name{rport.app.new}
\alias{rport.app.new}
\title{Creates an empty folder structure and file templates for an Rport app. If
directory `root/name` exists, stops with an error.}
\usage{
  rport.app.new(app.name, root.dir = getwd())
}
\arguments{
  \item{name}{name of the app}

  \item{name}{of the project}
}
\description{
  Creates an empty folder structure and file templates for
  an Rport app. If directory `root/name` exists, stops with
  an error.
}

