% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrappers.R
\name{rport.bootstrap.executable}
\alias{rport.bootstrap.executable}
\title{Source all files in `lib/projects/my_project/*.R` and `lib/opts/*.R`}
\usage{
rport.bootstrap.executable(project, shared = TRUE, opts = TRUE)
}
\arguments{
\item{shared}{should the .r or .R files from `lib/shared` be sourced.}

\item{opts}{should the .r or .R files from `lib/shared` be sourced.}
}
\description{
Source all files in `lib/projects/my_project/*.R` and `lib/opts/*.R`
}

