\name{rport.bootstrap}
\alias{rport.bootstrap}
\title{Bootstraps specific feature of an Rport app. Loads dependencies and sets up.}
\usage{
  rport.bootstrap(arg, ...)
}
\arguments{
  \item{arg}{what should we bootstrap: 'executable',
  'mailing', 'parallel'}
}
\description{
  Bootstraps specific feature of an Rport app. Loads
  dependencies and sets up.
}

