% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_cache.R
\name{rport.db.cache.clean}
\alias{rport.db.cache.clean}
\title{Remove all cache files from tmp/cache folder that have names not found in the
.Rportcache file}
\usage{
rport.db.cache.clean()
}
\description{
Remove all cache files from tmp/cache folder that have names not found in the
.Rportcache file
}

