% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_cache.R
\name{rport.db.cache.file.name}
\alias{rport.db.cache.file.name}
\title{Generate the file name of a cache file given query}
\usage{
rport.db.cache.file.name(query)
}
\arguments{
\item{string}{file name of the cache file}
}
\value{
string file name of the cache file
}
\description{
Generate the file name of a cache file given query
}

