\name{rport.db.cache.records}
\alias{rport.db.cache.records}
\title{If `hash` is not NULL returns TRUE/FALSE if `hash` exists in the cache
records. If `hash` is null returns the full list of records.}
\usage{
  rport.db.cache.records(query = NULL, save = FALSE)
}
\description{
  If `hash` is not NULL returns TRUE/FALSE if `hash` exists
  in the cache records. If `hash` is null returns the full
  list of records.
}

