\name{rport.db.set.env}
\alias{rport.db.set.env}
\title{This creates an RportDbEnvironment variable in the .GlobalEnv. This is so far
the only way to dynamically create objects from a package and should be used
with care as it violates the package Namespace.}
\usage{
  rport.db.set.env()
}
\description{
  This creates an RportDbEnvironment variable in the
  .GlobalEnv. This is so far the only way to dynamically
  create objects from a package and should be used with
  care as it violates the package Namespace.
}

