% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{rport.environment}
\alias{rport.environment}
\title{Return the name of Rport runtime environment.}
\usage{
rport.environment()
}
\value{
string name of environment
}
\description{
Return the name of Rport runtime environment.
}

