% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{rport.is.root.dir}
\alias{rport.is.root.dir}
\title{Is the given path a root directory of an Rport app}
\usage{
rport.is.root.dir(path)
}
\arguments{
\item{path}{character}
}
\value{
TRUE/FALSE
}
\description{
Is the given path a root directory of an Rport app
}
