\name{rport.new}
\alias{rport.new}
\title{Creates the file sceleton and codebase for a new Rport project. Use it to
start a new report, repeatable task, etc.}
\usage{
  rport.new(name)
}
\arguments{
  \item{env}{string specifying configuration settings to
  use.}
}
\description{
  Creates the file sceleton and codebase for a new Rport
  project. Use it to start a new report, repeatable task,
  etc.
}

