% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{rport.project.new}
\alias{rport.project.new}
\title{Creates a new project with templates.}
\usage{
rport.project.new(project.name, root.dir = getwd())
}
\arguments{
\item{name}{name of the app}

\item{name}{of the project}
}
\description{
Creates a new project with templates.
}
