\name{rport}
\alias{rport}
\title{Bootstraps your Rport project using the `env` configuration. Should be
placed at the beginning of each project. This is done automatically if you
use the project generators.}
\usage{
  rport(env)
}
\arguments{
  \item{env}{string specifying configuration settings to
  use.}
}
\description{
  Bootstraps your Rport project using the `env`
  configuration. Should be placed at the beginning of each
  project. This is done automatically if you use the
  project generators.
}

