% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rport.R
\name{rport.runtime.exit}
\alias{rport.runtime.exit}
\title{Prepares Rport for exist.}
\usage{
rport.runtime.exit()
}
\arguments{
\item{env}{string specifying configuration settings to use.}
}
\description{
* closes database connections
}

