% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{rport.specs.all}
\alias{rport.specs.all}
\title{Runs the specs of an app found in the app's spec/ folder.
You can use this function to run the specs from an interactive R session.
This is what is used by the specs/all executable.}
\usage{
rport.specs.all(root.dir = getwd())
}
\arguments{
\item{root.dir}{the path of the app}
}
\description{
Runs the specs of an app found in the app's spec/ folder.
You can use this function to run the specs from an interactive R session.
This is what is used by the specs/all executable.
}
