% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rport.R
\name{db.disconnect}
\alias{db.disconnect}
\title{Disconnect database connections. If `con.name` is not NA, then a connection
is closed by the given name. Otherwise all open database connections are
closed.}
\usage{
db.disconnect(con.name = NA)
}
\description{
Disconnect database connections. If `con.name` is not NA, then a connection
is closed by the given name. Otherwise all open database connections are
closed.
}
