% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrappers.R
\name{rport.bootstrap}
\alias{rport.bootstrap}
\title{Bootstraps specific feature of an Rport app. Loads dependencies and sets up.}
\usage{
rport.bootstrap(arg, ...)
}
\arguments{
\item{arg}{what should we bootstrap: 'executable', 'mailing', 'parallel'}
}
\description{
Bootstraps specific feature of an Rport app. Loads dependencies and sets up.
}

