\name{rport.db.cache.clean}
\alias{rport.db.cache.clean}
\title{Remove all cache files from tmp/cache folder that have names not found in the
.Rportcache file}
\usage{
  rport.db.cache.clean()
}
\description{
  Remove all cache files from tmp/cache folder that have
  names not found in the .Rportcache file
}

