\name{rport.db.cache.dir}
\alias{rport.db.cache.dir}
\title{TODO: for now hardcode this, but we could export this as a setting in
config/settings.R. The folder that will contain db cache files.}
\usage{
  rport.db.cache.dir()
}
\description{
  TODO: for now hardcode this, but we could export this as
  a setting in config/settings.R. The folder that will
  contain db cache files.
}

