\name{rport.environment.is}
\alias{rport.environment.is}
\title{Check if the given string is the current runtime environment of Rport.
Useful to execute code conditional on runtime environment, e.g. only
development or only production.}
\usage{
  rport.environment.is(env)
}
\arguments{
  \item{env}{string the name of the environment to compare}
}
\value{
  TRUE/FALSE
}
\description{
  Check if the given string is the current runtime
  environment of Rport. Useful to execute code conditional
  on runtime environment, e.g. only development or only
  production.
}

