% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rport.R
\name{rport.root}
\alias{rport.root}
\title{The root directory of the Rport app. For now consider this to be the working
directory of the R process.}
\usage{
rport.root()
}
\description{
The root directory of the Rport app. For now consider this to be the working
directory of the R process.
}

